messenger/requirements.txt
astroid==2.15.3
autopep8==2.0.2
dill==0.3.6
greenlet==2.0.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.2.0
pycodestyle==2.10.0
pylint==2.17.2
python-dotenv==1.0.0
PyYAML==6.0
SQLAlchemy==1.4.0
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
wrapt==1.15.0
